llm_qa/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.0
azure-search==1.0.0b2
azure-search-documents==11.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
executing==2.0.1
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
jedi==0.19.1
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.25
marshmallow==3.20.2
matplotlib-inline==0.1.6
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.11.1
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
parso==0.8.3
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PySocks==1.7.1
pytest==8.0.1
pytest-asyncio==0.23.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==25.1.2
rank-bm25==0.2.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
tiktoken==0.6.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.2
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
yarl==1.9.4
